separate liveness and startup checks
