+++ BookingsApi/BookingsApi/Startup.cs
@@ -12,6 +12,7 @@ using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.Extensions.Hosting;
 using BookingsApi.Contract.V1.Configuration;
 using BookingsApi.Domain.Configuration;
+using BookingsApi.Health;
 using BookingsApi.Validations.Common;
 using FluentValidation;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
@@ -165,28 +166,35 @@ namespace BookingsApi
             {
                 endpoints.MapDefaultControllerRoute();
                 
+                // TODO: need to update the config. currently this route is used for liveness and readiness checks
                 endpoints.MapHealthChecks("/healthcheck/health", new HealthCheckOptions()
                 {
-                    ResponseWriter = async (context, report) =>
-                    {
-                        var result = JsonConvert.SerializeObject(
-                            new
-                            {
-                                status = report.Status.ToString(),
-                                details = report.Entries.Select(e => new
-                                {
-                                    key = e.Key, 
-                                    value = Enum.GetName(typeof(HealthStatus), e.Value.Status),
-                                    error = e.Value.Exception?.Message
-                                })
-                            });
-                        context.Response.ContentType = "application/json";
-                        await context.Response.WriteAsync(result);
-                    }
+                    Predicate = check => check.Tags.Contains("self"),
+                    ResponseWriter = HealthCheckResponseWriter
                 });
 
-                endpoints.MapHealthChecks("health/liveness");
+                // TODO: need to update the config. currently the liveness route is used for startup
+                endpoints.MapHealthChecks("health/liveness", new HealthCheckOptions()
+                {
+                    Predicate = check => check.Tags.Contains("services"),
+                    ResponseWriter = HealthCheckResponseWriter
+                });
             });
+        }
+        
+        private async Task HealthCheckResponseWriter(HttpContext context, HealthReport report)
+        {
+            var result = JsonConvert.SerializeObject(new
+            {
+                status = report.Status.ToString(),
+                details = report.Entries.Select(e => new
+                {
+                    key = e.Key, value = Enum.GetName(typeof(HealthStatus), e.Value.Status),
+                    error = e.Value.Exception?.Message
+                })
+            });
+            context.Response.ContentType = "application/json";
+            await context.Response.WriteAsync(result);
         }
     }
 }
